[*] [Implementation] Fix to work with Arduino..
Solved that mystery that increasing wave frequency increases time spent
in main cycle with interrupts.

Arduino uses counter 1 as timer. We're using counter 3 for that.
And we were ignoring Arduino because we don't use it's millis().

But counter 1 has interrupt on overflow. Which calls Arduino's code.
And looks like that overflow occurs every time our half wave is done.
(Which can be up to 2 MHz or so.)

So yeah, we're disabling interrupt on overflow. Which for single-slope
occurs every time we're reaching Mark A.